+++ sources/Device/src/Tester/Tester.h
@@ -43,6 +43,6 @@ namespace Tester
     public:
         static void Update();
         /// Устанавливает точки для рисования, соответствующие шагу numStep
-        static void SetPoints(int numStep, const uint8 dx[TESTER_NUM_POINTS], const uint8 dy[TESTER_NUM_POINTS]);
+        static void SetPoints(int numStep, const uint16 dx[TESTER_NUM_POINTS], const uint8 dy[TESTER_NUM_POINTS]);
     };
 };
